JBackgroundPanel and TitleScreen are ones that I've created to see if I can get it to work. It doesn't. Wanted to see if anyone else can maybe figure it out. If not I'll delete it tomorrow.
My failed attempt and me saying I'm not gonna try to work on this anymore tonight because I'm not exhausted and frustrated and feeling nauseated. >.< Can someone please please please try and figure it out? I made the image in hopes that it would make our game more unified in the theme.